Moved fakes under main implementation
