+++ pkg/server/profiles_test.go
@@ -19,8 +19,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	pb "github.com/weaveworks/weave-gitops/pkg/api/profiles"
-	"github.com/weaveworks/weave-gitops/pkg/fakes"
 	"github.com/weaveworks/weave-gitops/pkg/helm/watcher/cache/cachefakes"
+	"github.com/weaveworks/weave-gitops/pkg/kube/kubefakes"
 	"github.com/weaveworks/weave-gitops/pkg/server"
 	"github.com/weaveworks/weave-gitops/pkg/testutils"
 )
@@ -45,7 +45,7 @@ var _ = Describe("ProfilesServer", func() {
 		kubeClient = fake.NewClientBuilder().WithScheme(scheme).Build()
 
 		fakeCache = &cachefakes.FakeCache{}
-		fakeClientGetter := fakes.NewFakeClientGetter(kubeClient)
+		fakeClientGetter := kubefakes.NewFakeClientGetter(kubeClient)
 		s = &server.ProfilesServer{
 			Log:               testutils.MakeFakeLogr(),
 			HelmRepoName:      "helmrepo",
